refactor: マイページリファクタリング #175
