+++ front/src/app/my-page/page.tsx
@@ -1,5 +1,5 @@
 import { getUser } from "@/lib/user";
-import MyPagePresentation from "../presentation/MyPage/MyPagePresentation";
+import MyPagePresentation from "./MyPagePresentation";
 
 export default async function MyPage() {
   const user = await getUser();
