+++ front/src/features/my-page/DeleteForm/DeleteForm.module.scss
@@ -1,36 +1,3 @@
-.wrapper {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: calc(100vh - var(--header-height));
-  position: relative;
-  z-index: 10;
-}
-
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 40px;
-  width: 90%;
-  border: 1px solid #747775;
-  border-radius: 10px;
-  background-color: #fff;
-
-  @media (min-width: 768px) {
-    max-width: 560px;
-  }
-}
-
-.content {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
 .form {
   width: 100%;
   height: 100%;
@@ -38,14 +5,6 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
-}
-
-.title {
-  text-align: center;
-  color: var(--black);
-  font-size: 20px;
-  font-weight: 700;
-  margin-bottom: 2rem;
 }
 
 .accountContainer {
